move lsp config into a folder
